added files, reformat and 1 unnecess. malloc down
